- Started on multiplechoice

+++ app/game/events.py
@@ -8,6 +8,9 @@ from app.extensions import db, socketio
 
 from .models import Room
 
+import requests
+import time
+
 
 @socketio.on('join_game', namespace='/game')
 def join_game(data):
@@ -16,16 +19,19 @@ def join_game(data):
     join_room(room_id)
 
     room = Room.query.filter_by(room_id=room_id).first()
-    
-    players = []
 
-    for player in room.players:
-        players.append({
-            'user_id' : player.id,
-            'username' : player.username
-        })
+    if room:
+        players = []
 
-    emit('update_players', {'owner_id': room.owner_id, 'players': players}, room=room_id, namespace='/game', broadcast=True)
+        for player in room.players:
+            players.append({
+                'user_id' : player.id,
+                'username' : player.username
+            })
+
+        emit('update_players', {'owner_id': room.owner_id, 'players': players}, room=room_id, namespace='/game', broadcast=True)
+    else:
+        emit('kick_all', room=room_id, namespace='/game', broadcast=True)
 
 
 @socketio.on('leave_game', namespace='/game')
@@ -36,19 +42,42 @@ def leave_game(data):
 
     room = Room.query.filter_by(room_id=room_id).first()
 
-    room.players.remove(current_user)
+    if room:
+        room.players.remove(current_user)
 
-    if current_user.id == room.owner_id: # type: ignore
-        db.session.delete(room)
+        if current_user.id == room.owner_id: # type: ignore
+            db.session.delete(room)
 
-    db.session.commit()
+        db.session.commit()
 
-    players = []
+        players = []
 
-    for player in room.players:
-        players.append({
-            'user_id' : player.id,
-            'username' : player.username
-        })
+        for player in room.players:
+            players.append({
+                'user_id' : player.id,
+                'username' : player.username
+            })
 
-    emit('update_players', {'owner_id': room.owner_id, 'players': players}, room=room_id, namespace='/game', broadcast=True)
+        emit('update_players', {'owner_id': room.owner_id, 'players': players}, room=room_id, namespace='/game', broadcast=True)
+    else:
+        emit('kick_all', room=room_id, namespace='/game', broadcast=True)
+
+
+@socketio.on('start_game', namespace='/game')
+def start_game(data):
+    room_id = data['room_id']
+
+    emit('starting', room=room_id, namespace='/game', broadcast=True)
+
+    time.sleep(5)
+
+    room = Room.query.filter_by(room_id=room_id).first()
+
+    room.started = True
+
+    response = requests.get('https://opentdb.com/api.php?amount=1&difficulty=medium&type=multiple')
+
+    print()
+
+    if response.ok:
+        emit('question', response.json()['results'][0], room=room_id, namespace='/game', broadcast=True)
